feat: autogpt-as-a-service
